resources/requirements.txt
# -------------------------
# runtime
# -------------------------
customtkinter==5.2.2     # GUI moderna sobre Tkinter :contentReference[oaicite:0]{index=0}
pyodbc==5.2.0            # Conexión ODBC a SQL Server, etc. :contentReference[oaicite:1]{index=1}
Pillow==11.3.0           # Manejo de imágenes, requerido por customtkinter :contentReference[oaicite:2]{index=2}

# -------------------------
# build-only (opcional)
# -------------------------
pyinstaller==6.14.2      # Generar el .exe en modo onefile :contentReference[oaicite:3]{index=3}
auto-py-to-exe==2.46.0   # GUI para PyInstaller (facilita el proceso) :contentReference[oaicite:4]{index=4}
